Add files via upload
This model uses a Random Forest Classifier to predict the outcome or type of a game based on features from the Games.csv dataset. The model is trained after handling missing values and encoding categorical data. It provides accurate predictions and helps understand which game attributes influence outcomes the most.